changes made so that mobile ios will build and run normally for the most part.
