+++ ios/Runner/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyD9GT7aIZ55zPhz25WflfuRsRMOYqDqog4</string>
+	<key>GCM_SENDER_ID</key>
+	<string>635469629055</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>app.intrst</string>
+	<key>PROJECT_ID</key>
+	<string>intrst-1525412155568</string>
+	<key>STORAGE_BUCKET</key>
+	<string>intrst-1525412155568.firebasestorage.app</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:635469629055:ios:fdc6bd2272ee754e6fe14a</string>
+</dict>
+</plist>
