+++ lib/utility/GeneralUtilityStub.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'GeneralUtility.dart';
+
+class GeneralUtilityWeb implements GeneralUtility {
+  @override
+  bool isMobileBrowser(BuildContext context) {
+  return true;
+  }
+}
